Resturants_template
Lambda is modern and creative one-page fully responsive HTML5 template. It is suitable for any restaurant, a small restaurant or a premium that would briefly talk about his restaurant and show all the achievements, menus and restaurant interiors. Template is designed to comply with these requirements and all the latest trends of web design.